фиксы

+++ src/main/java/com/cyanide3d/lib/mylittleorm/proxy/configurators/DeleteMethodConfigurer.java
@@ -1,5 +1,6 @@
 package com.cyanide3d.lib.mylittleorm.proxy.configurators;
 
+import com.cyanide3d.exception.UnsupportedReturnTypeException;
 import com.cyanide3d.lib.mylittleorm.handler.DatabaseProvider;
 
 import java.lang.reflect.Method;
@@ -14,6 +15,10 @@ public class DeleteMethodConfigurer implements MethodConfigurer{
 
     @Override
     public Object configure(Method method, Object[] args, Class<?> clazz) {
+        String returnTypeName = method.getReturnType().getName();
+        if (!returnTypeName.equalsIgnoreCase("void")) {
+            throw new UnsupportedReturnTypeException("Method " + method.getName() + " can't use return type [" + returnTypeName + "].");
+        }
         if (method.getName().equalsIgnoreCase("delete")) {
             dao.delete(args[0]);
             return true;
